backend/ocr_service/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.0.0

# Image Processing
Pillow==10.1.0

# HTTP Client
requests>=2.31.0
charset-normalizer>=3.0.0

# Utilities
python-dotenv==1.0.0
